Make "Advanced Unit Selection" dialog movable
