+++ client/dialogs.h
@@ -84,21 +84,6 @@ protected:
 };
 
 /***************************************************************************
-  Nonmodal message box for disbanding units
-***************************************************************************/
-class disband_box : public hud_message_box {
-  Q_OBJECT
-  const std::vector<unit *> cpunits;
-
-public:
-  explicit disband_box(const std::vector<unit *> &punits,
-                       QWidget *parent = 0);
-  ~disband_box() override;
-private slots:
-  void disband_clicked();
-};
-
-/***************************************************************************
  Dialog for goto popup
 ***************************************************************************/
 class notify_goto : public QMessageBox {
